requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
django-filter
twilio
python-decouple
drf-yasg
